fix: update project overview
