+++ frontend/src/component/project/Project/ProjectStats/StatusBox.tsx
@@ -1,10 +1,11 @@
-import { ArrowOutward, SouthEast } from '@mui/icons-material';
+import type { ReactNode } from 'react';
+import { CallMade, SouthEast } from '@mui/icons-material';
 import { Box, Typography, styled } from '@mui/material';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import { flexRow } from 'themes/themeStyles';
 
 const StyledBox = styled(Box)(({ theme }) => ({
-    padding: theme.spacing(4, 2),
+    padding: theme.spacing(3),
     backgroundColor: theme.palette.background.paper,
     minWidth: '24%',
     display: 'flex',
@@ -26,12 +27,11 @@ const StyledBox = styled(Box)(({ theme }) => ({
 }));
 
 const StyledTypographyHeader = styled(Typography)(({ theme }) => ({
-    marginBottom: theme.spacing(2),
+    marginBottom: theme.spacing(2.5),
 }));
 
 const StyledTypographyCount = styled(Typography)(({ theme }) => ({
     fontSize: theme.fontSizes.largeHeader,
-    fontWeight: 'bold',
 }));
 
 const StyledBoxChangeContainer = styled(Box)(({ theme }) => ({
@@ -43,17 +43,18 @@ const StyledBoxChangeContainer = styled(Box)(({ theme }) => ({
 
 const StyledTypographySubtext = styled(Typography)(({ theme }) => ({
     color: theme.palette.neutral.main,
-    fontSize: theme.fontSizes.smallBody,
+    fontSize: theme.typography.body2.fontSize,
 }));
 
 const StyledTypographyChange = styled(Typography)(({ theme }) => ({
     marginLeft: theme.spacing(1),
-    fontSize: theme.fontSizes.smallBody,
+    fontSize: theme.typography.body1.fontSize,
+    fontWeight: theme.typography.fontWeightBold,
 }));
 
 interface IStatusBoxProps {
     title: string;
-    boxText: string;
+    boxText: ReactNode;
     change: number;
     percentage?: boolean;
     fullWidthBodyText?: boolean;
@@ -62,19 +63,19 @@ interface IStatusBoxProps {
 const resolveIcon = (change: number) => {
     if (change > 0) {
         return (
-            <ArrowOutward
-                sx={{ color: 'success.main', height: 18, width: 18 }}
+            <CallMade
+                sx={{ color: 'success.dark', height: 20, width: 20 }}
             />
         );
     }
-    return <SouthEast sx={{ color: 'warning.dark', height: 18, width: 18 }} />;
+    return <SouthEast sx={{ color: 'warning.dark', height: 20, width: 20 }} />;
 };
 
 const resolveColor = (change: number) => {
     if (change > 0) {
-        return 'success.main';
+        return 'success.dark';
     }
-    return 'error.main';
+    return 'warning.dark';
 };
 
 export const StatusBox = ({
@@ -110,6 +111,7 @@ export const StatusBox = ({
                                 <StyledTypographyChange
                                     color={resolveColor(change)}
                                 >
+                                    {change > 0 ? '+' : ''}
                                     {change}
                                     {percentage ? '%' : ''}
                                 </StyledTypographyChange>
